Fix Mobile Nav Bug
Although I did my best to set z-indexes properly for mobile navbar's elements, it still wasn't displaying properly - when nav-items were supposed to slide up under the navbar, they were showing up on the surface. I didn't want to rewrite the whole navbar, so I decided to change the animation, so the nav-items are sliding in and out from the side.